+++ example/lib/config/agora.config.dart
@@ -1,14 +1,31 @@
 /// Get your own App ID at https://dashboard.agora.io/
-const appId = YOUR_APP_ID;
+String get appId {
+  // Allow pass an `appId` as an environment variable with name `TEST_APP_ID` by using --dart-define
+  return const String.fromEnvironment('TEST_APP_ID',
+      defaultValue: '<YOUR_APP_ID>');
+}
 
 /// Please refer to https://docs.agora.io/en/Agora%20Platform/token
-const token = YOUR_TOEKN;
+String get token {
+  // Allow pass a `token` as an environment variable with name `TEST_TOKEN` by using --dart-define
+  return const String.fromEnvironment('TEST_TOKEN',
+      defaultValue: '<YOUR_TOKEN>');
+}
 
 /// Your channel ID
-const channelId = YOUR_CHANNEL_ID;
+String get channelId {
+  // Allow pass a `channelId` as an environment variable with name `TEST_CHANNEL_ID` by using --dart-define
+  return const String.fromEnvironment(
+    'TEST_CHANNEL_ID',
+    defaultValue: '<YOUR_CHANNEL_ID>',
+  );
+}
 
 /// Your int user ID
-const uid = YOUR_UID;
+const int uid = 0;
+
+/// Your user ID for the screen sharing
+const int screenSharingUid = 10;
 
 /// Your string user ID
-const stringUid = YOUR_STRING_UID;
+const String stringUid = '0';