+++ example/lib/examples/basic/index.dart
@@ -1,13 +1,11 @@
-import 'package:agora_rtc_engine_example/examples/basic/join_channel_audio.dart';
-import 'package:agora_rtc_engine_example/examples/basic/join_channel_external_audio.dart';
-import 'package:agora_rtc_engine_example/examples/basic/join_channel_video.dart';
-import 'package:agora_rtc_engine_example/examples/basic/string_uid.dart';
+import 'package:agora_rtc_engine_example/examples/basic/join_channel_audio/join_channel_audio.dart';
+import 'package:agora_rtc_engine_example/examples/basic/join_channel_video/join_channel_video.dart';
+import 'package:agora_rtc_engine_example/examples/basic/string_uid/string_uid.dart';
 
 /// Data source for basic examples
-final Basic = [
+final basic = [
   {'name': 'Basic'},
-  {'name': 'JoinChannelAudio', 'widget': JoinChannelAudio()},
-  {'name': 'JoinChannelExternalAudio', 'widget': JoinChannelExternalAudio()},
-  {'name': 'JoinChannelVideo', 'widget': JoinChannelVideo()},
-  {'name': 'StringUid', 'widget': StringUid()}
+  {'name': 'JoinChannelAudio', 'widget': const JoinChannelAudio()},
+  {'name': 'JoinChannelVideo', 'widget': const JoinChannelVideo()},
+  {'name': 'StringUid', 'widget': const StringUid()}
 ];